% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp-gan-trainer.R
\name{calibrate_noise_multiplier}
\alias{calibrate_noise_multiplier}
\title{Calibrate Noise Multiplier for Target Privacy}
\usage{
calibrate_noise_multiplier(
  target_epsilon,
  target_delta,
  sampling_rate,
  total_steps,
  min_noise = 0.1,
  max_noise = 100,
  tolerance = 0.01
)
}
\arguments{
\item{target_epsilon}{Target epsilon for differential privacy}

\item{target_delta}{Target delta for differential privacy}

\item{sampling_rate}{Poisson sampling probability}

\item{total_steps}{Total number of training steps}

\item{min_noise}{Minimum noise multiplier to search. Defaults to 0.1.}

\item{max_noise}{Maximum noise multiplier to search. Defaults to 100.}

\item{tolerance}{Convergence tolerance. Defaults to 0.01.}
}
\value{
Calibrated noise multiplier
}
\description{
Find the noise multiplier that achieves a target epsilon
for a given number of training steps using binary search.
}
\keyword{internal}
