% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_discriminator.R
\name{create_discriminator}
\alias{create_discriminator}
\title{Create a GAN Discriminator}
\usage{
create_discriminator(a = "Hello", b = "World")
}
\arguments{
\item{a}{Input a}

\item{b}{Input b}
}
\value{

}
\description{
Create a GAN discriminator. You can either input your own architecture or rely on the default definition.
}
\examples{
create_discriminator(a = "A", b = "B")
}
