% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataToTorch.R
\name{DataToTorch}
\alias{DataToTorch}
\title{DataToTorch}
\usage{
DataToTorch(transformed_data, device = "cpu")
}
\arguments{
\item{transformed_data}{Input a data set after DataTransformer}

\item{device}{Input on which device (e.g. "cpu" or "cuda") will you be training?}
}
\value{
A function
}
\description{
Provides a function to send the output of a DataTransformer to
a torch tensor, so that it can be accessed during GAN training.
}
