% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discriminator.R
\name{DCGAN_Discriminator}
\alias{DCGAN_Discriminator}
\title{DCGAN Discriminator}
\usage{
DCGAN_Discriminator(
  image_size = 64,
  number_channels = 3,
  ndf = 64,
  dropout_rate = 0.5,
  sigmoid = FALSE
)
}
\value{
A neural net for the DCGAN Discriminator
}
\description{
Provides a class to transform data for RGAN.
Method \verb{$new()} initializes a new transformer, method \verb{$fit(data)} learns
the parameters for the transformation from data (e.g. means and sds).
Methods \verb{$transform()} and \verb{$inverse_transform()} can be used to transform
and back transform a data set based on the learned parameters.
Currently, DataTransformer supports z-transformation (a.k.a. normalization)
for numerical features/variables and one hot encoding for categorical
features/variables. In your call to fit you just need to indicate which
columns contain discrete features.
}
