% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generator.R
\name{DCGAN_Generator}
\alias{DCGAN_Generator}
\title{DCGAN Generator}
\usage{
DCGAN_Generator(
  noise_dim = 100,
  image_size = 64,
  number_channels = 3,
  ngf = 64
)
}
\arguments{
\item{noise_dim}{a}

\item{data_dim}{b}

\item{hidden_units}{list of number of neurons per layer the length of the list determines the number of hidden layers}

\item{dropout_rate}{dropout for each hidden layer}
}
\value{
A neural net class for the DCGAN Generator
}
\description{
Provides a class to transform data for RGAN.
Method \verb{$new()} initializes a new transformer, method \verb{$fit(data)} learns
the parameters for the transformation from data (e.g. means and sds).
Methods \verb{$transform()} and \verb{$inverse_transform()} can be used to transform
and back transform a data set based on the learned parameters.
Currently, DataTransformer supports z-transformation (a.k.a. normalization)
for numerical features/variables and one hot encoding for categorical
features/variables. In your call to fit you just need to indicate which
columns contain discrete features.
Architecture from: https://pytorch.org/tutorials/beginner/dcgan_faces_tutorial.html
}
