% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminator-architectures.R
\name{Discriminator}
\alias{Discriminator}
\title{Discriminator}
\usage{
Discriminator(
  data_dim,
  hidden_units = list(128, 128),
  dropout_rate = 0.5,
  sigmoid = FALSE
)
}
\arguments{
\item{data_dim}{The number of columns in the data set}

\item{hidden_units}{A list of the number of neurons per layer, the length of the list determines the number of hidden layers}

\item{dropout_rate}{The dropout rate for each hidden layer}

\item{sigmoid}{Switch between a sigmoid and linear output layer (the sigmoid is needed for the original GAN value function)}
}
\value{
A torch::nn_module for the Discriminator
}
\description{
Provides a torch::nn_module with a simple fully connected neural
net, for use as the default architecture for tabular data in RGAN.
}
