% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp-gan-trainer.R
\name{dp_accountant_poisson}
\alias{dp_accountant_poisson}
\title{RDP Privacy Accountant for Poisson Subsampling}
\description{
R6 class for tracking privacy loss using Renyi Differential Privacy
with Poisson subsampling amplification.
}
\details{
Implements the privacy accounting from Mironov (2017) with subsampling
amplification. RDP provides tighter composition than basic composition
and converts to (epsilon, delta)-DP guarantees.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{steps}}{Number of steps taken}

\item{\code{sampling_rate}}{Poisson sampling rate}

\item{\code{noise_multiplier}}{Gaussian noise multiplier}

\item{\code{target_delta}}{Target delta for conversion}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-dp_accountant_poisson-new}{\code{dp_accountant_poisson$new()}}
\item \href{#method-dp_accountant_poisson-step}{\code{dp_accountant_poisson$step()}}
\item \href{#method-dp_accountant_poisson-compute_rdp}{\code{dp_accountant_poisson$compute_rdp()}}
\item \href{#method-dp_accountant_poisson-get_epsilon}{\code{dp_accountant_poisson$get_epsilon()}}
\item \href{#method-dp_accountant_poisson-clone}{\code{dp_accountant_poisson$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dp_accountant_poisson-new"></a>}}
\if{latex}{\out{\hypertarget{method-dp_accountant_poisson-new}{}}}
\subsection{Method \code{new()}}{
Create a new privacy accountant
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dp_accountant_poisson$new(
  sampling_rate,
  noise_multiplier,
  target_delta = 1e-05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sampling_rate}}{Poisson sampling probability}

\item{\code{noise_multiplier}}{Noise scale relative to sensitivity}

\item{\code{target_delta}}{Target delta for DP guarantee}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dp_accountant_poisson-step"></a>}}
\if{latex}{\out{\hypertarget{method-dp_accountant_poisson-step}{}}}
\subsection{Method \code{step()}}{
Record one or more privacy-consuming steps
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dp_accountant_poisson$step(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of steps to record}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dp_accountant_poisson-compute_rdp"></a>}}
\if{latex}{\out{\hypertarget{method-dp_accountant_poisson-compute_rdp}{}}}
\subsection{Method \code{compute_rdp()}}{
Compute RDP at a given order
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dp_accountant_poisson$compute_rdp(alpha)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{RDP order (must be > 1)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
RDP value at order alpha for one step
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dp_accountant_poisson-get_epsilon"></a>}}
\if{latex}{\out{\hypertarget{method-dp_accountant_poisson-get_epsilon}{}}}
\subsection{Method \code{get_epsilon()}}{
Get (epsilon, delta)-DP guarantee
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dp_accountant_poisson$get_epsilon(delta = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{Target delta (defaults to target_delta)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Epsilon value for the given delta
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dp_accountant_poisson-clone"></a>}}
\if{latex}{\out{\hypertarget{method-dp_accountant_poisson-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dp_accountant_poisson$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
