% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-synthetic-data.R
\name{expert_sample_synthetic_data}
\alias{expert_sample_synthetic_data}
\title{Sample Synthetic Data with explicit noise input}
\usage{
expert_sample_synthetic_data(g_net, z, device, eval_dropout = FALSE)
}
\arguments{
\item{g_net}{A torch::nn_module with a Generator}

\item{z}{A noise vector}

\item{device}{The device on which synthetic data should be sampled (cpu or cuda)}

\item{eval_dropout}{Should dropout be applied during inference}
}
\value{
Synthetic data
}
\description{
Provides a function that makes it easy to sample synthetic data from a Generator
}
