% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp-gan-trainer.R
\name{fast_poisson_subsample}
\alias{fast_poisson_subsample}
\title{Fast Poisson Subsampling (Non-Cryptographic)}
\usage{
fast_poisson_subsample(n_samples, sampling_rate)
}
\arguments{
\item{n_samples}{Total number of samples in the dataset}

\item{sampling_rate}{Probability of including each sample}
}
\value{
Integer vector of selected indices
}
\description{
Fast Poisson subsampling using R's standard RNG.
Not cryptographically secure but much faster for development/testing.
}
\keyword{internal}
