% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gan-fns.R
\name{gan}
\alias{gan}
\title{GAN}
\usage{
gan(data, epochs = 500, batch_size = 32, model_name = "gan_r",
  continue_training = FALSE, trace = 100)
}
\arguments{
\item{batch_size}{}
}
\description{
Runs the GAN in Python and returns the results
}
