% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gan-trainer.R
\name{GAN_update_plot_image}
\alias{GAN_update_plot_image}
\title{GAN_update_plot_image}
\usage{
GAN_update_plot_image(mfrow = c(4, 4), synth_data)
}
\arguments{
\item{mfrow}{The dimensions of the grid of images to be plotted}

\item{synth_data}{The synthetic data (images) to be plotted}
}
\value{
A function
}
\description{
Provides a function to send the output of a DataTransformer to
a torch tensor, so that it can be accessed during GAN training.
}
