% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GANTrainer.R
\name{GAN_update_step}
\alias{GAN_update_step}
\title{GAN_update_step}
\usage{
GAN_update_step(
  data,
  batch_size,
  device = "cpu",
  g_net,
  d_net,
  g_optimizer,
  d_optimizer,
  value_function,
  weight_clipper
)
}
\arguments{
\item{device}{Input on which device (e.g. "cpu" or "cuda") will you be training?}

\item{transformed_data}{Input a data set after DataTransformer}
}
\value{
A function
}
\description{
Provides a function to send the output of a DataTransformer to
a torch tensor, so that it can be accessed during GAN training.
}
