% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value-functions.R
\name{GAN_value_fct}
\alias{GAN_value_fct}
\title{GAN_value_fct}
\usage{
GAN_value_fct(real_scores, fake_scores)
}
\arguments{
\item{transformed_data}{Input a data set after DataTransformer}

\item{device}{Input on which device (e.g. "cpu" or "cuda") will you be training?}
}
\value{
A function
}
\description{
Provides a function to send the output of a DataTransformer to
a torch tensor, so that it can be accessed during GAN training.
}
