% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp-gan-trainer.R
\name{generator_step}
\alias{generator_step}
\title{Generator Update Step}
\usage{
generator_step(
  batch_size,
  noise_dim,
  sample_noise,
  device,
  g_net,
  d_net,
  g_optim
)
}
\arguments{
\item{batch_size}{Number of samples to generate}

\item{noise_dim}{Dimension of generator noise}

\item{sample_noise}{Function to sample noise}

\item{device}{Computation device}

\item{g_net}{Generator network}

\item{d_net}{Discriminator network}

\item{g_optim}{Generator optimizer}
}
\value{
Scalar loss value
}
\description{
Standard (non-private) generator update step.
The generator doesn't need DP since it only accesses synthetic data.
}
\keyword{internal}
