% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gan-fns.R
\name{get_training_data}
\alias{get_training_data}
\title{Training data}
\usage{
get_training_data(n = 200, m = 250, shape = list(a = c(1, 3), p =
  c(2, 10)))
}
\arguments{
\item{n}{Number of waves}

\item{m}{Number of points to evaluate the sine waves}

\item{shape}{List of shape parameters for the two sine waves. \code{a} amplitude; \code{p} period}
}
\description{
Generates training data for testing the GAN. Function generates two sine waves with different periods and amplitudes
}
