% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{kl_gen}
\alias{kl_gen}
\title{KL WGAN loss for Generator training}
\usage{
kl_gen(dis_fake)
}
\arguments{
\item{dis_fake}{Discriminator scores on fake examples ($D(G(z))$)}
}
\value{
The loss
}
\description{
Utility function for the kl WGAN loss for Generator training as described in \url{https://arxiv.org/abs/1910.09779}
and implemented in python in \url{https://github.com/ermongroup/f-wgan}.
}
