% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{kl_real}
\alias{kl_real}
\title{KL WGAN loss on real examples}
\usage{
kl_real(dis_real)
}
\arguments{
\item{dis_real}{Discriminator scores on real examples ($D(x)$)}
}
\value{
The loss
}
\description{
Utility function for the kl WGAN loss for real examples as described in \url{https://arxiv.org/abs/1910.09779}
and implemented in python in \url{https://github.com/ermongroup/f-wgan}.
}
