% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value-functions.R
\name{KLWGAN_value_fct}
\alias{KLWGAN_value_fct}
\title{KLWGAN Value Function}
\usage{
KLWGAN_value_fct(real_scores, fake_scores)
}
\arguments{
\item{real_scores}{The discriminator scores on real examples ($D(x)$)}

\item{fake_scores}{The discriminator scores on fake examples ($D(G(z))$)}
}
\value{
The function returns a named list with the entries d_loss and g_loss
}
\description{
Provides a function to send the output of a DataTransformer to
a torch tensor, so that it can be accessed during GAN training.
}
