% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gan-trainer.R
\name{pack_samples}
\alias{pack_samples}
\title{Pack Samples for PacGAN}
\usage{
pack_samples(data, pac)
}
\arguments{
\item{data}{A torch tensor of shape (batch_size, data_dim)}

\item{pac}{Number of samples to pack together. batch_size must be divisible by pac.}
}
\value{
A torch tensor of shape (batch_size/pac, data_dim*pac)
}
\description{
Reshapes a batch of samples for PacGAN by concatenating \code{pac}
consecutive samples along the feature dimension. This allows the discriminator
to see multiple samples at once, helping it detect mode collapse.
}
\keyword{internal}
