% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gan-fns.R
\name{plot_curves}
\alias{plot_curves}
\title{Plot a set of curves}
\usage{
plot_curves(data, id = 1:2, nrow = NULL, k = NULL)
}
\arguments{
\item{data}{x_train}

\item{id}{which rows to plot. Defaults to 1:2}

\item{k}{}
}
\description{
For inspecting the output
}
