% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{sample_toydata}
\alias{sample_toydata}
\title{Sample Toydata}
\usage{
sample_toydata(n = 1000, sd = 0.3, seed = 20211111)
}
\arguments{
\item{n}{Number of observations to generate}

\item{sd}{Standard deviation of the normal distribution to generate y}

\item{seed}{A seed for the pseudo random number generator}
}
\value{
A matrix with two columns x and y
}
\description{
Sample Toydata to reproduce the examples in the paper.
}
