% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{torch_rand_ab}
\alias{torch_rand_ab}
\title{Uniform Random numbers between values a and b}
\usage{
torch_rand_ab(shape, a = -1, b = 1, ...)
}
\arguments{
\item{shape}{Vector of dimensions of resulting tensor}

\item{a}{Lower bound of uniform distribution to sample from}

\item{b}{Upper bound of uniform distribution to sample from}

\item{...}{Potential additional arguments}
}
\value{
A sample from the specified uniform distribution in a tensor with the specified shape
}
\description{
Provides a function to sample torch tensors from an arbitrary uniform distribution.
}
