% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value-functions.R
\name{WGAN_value_fct}
\alias{WGAN_value_fct}
\title{WGAN Value Function}
\usage{
WGAN_value_fct(real_scores, fake_scores)
}
\arguments{
\item{real_scores}{The discriminator scores on real examples ($D(x)$)}

\item{fake_scores}{The discriminator scores on fake examples ($D(G(z))$)}
}
\value{
The function returns a named list with the entries d_loss and g_loss
}
\description{
Implements the Wasserstein GAN (WGAN) value function as a function to be called in gan_trainer.
Note that for this to work properly you also need to implement a weight clipper (or other procedure) to constrain the Discriminator weights.
}
