% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gan-fns.R
\name{y_joint}
\alias{y_joint}
\title{Joint function}
\usage{
y_joint(m = 200, shape = list(a = c(1, 3), p = c(2, 10)))
}
\arguments{
\item{m}{Number of points to evaluate the function. Default 200}

\item{shape}{Shape of the sine waves. Default \code{list(a = c(1, 3), p = c(2, 10))}}
}
\description{
Joint function
}
