% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gan-trainer.R
\name{GAN_update_plot}
\alias{GAN_update_plot}
\title{GAN_update_plot}
\usage{
GAN_update_plot(data, dimensions = c(1, 2), synth_data, epoch, main = NULL)
}
\arguments{
\item{data}{Real data to be plotted}

\item{dimensions}{Which columns of the data should be plotted}

\item{synth_data}{The synthetic data to be plotted}

\item{epoch}{The epoch during training for the plot title}

\item{main}{An optional plot title}
}
\value{
A function
}
\description{
Provides a function to send the output of a DataTransformer to
a torch tensor, so that it can be accessed during GAN training.
}
