% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value-functions.R
\name{GAN_value_fct}
\alias{GAN_value_fct}
\title{GAN Value Function}
\usage{
GAN_value_fct(real_scores, fake_scores)
}
\arguments{
\item{real_scores}{The discriminator scores on real examples ($D(x)$)}

\item{fake_scores}{The discriminator scores on fake examples ($D(G(z))$)}
}
\value{
The function returns a named list with the entries d_loss and g_loss
}
\description{
Implements the original GAN value function as a function to be called in gan_trainer.
The function can serve as a template to implement new value functions in RGAN.
}
