% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GANTrainer.R
\name{GANTrainer}
\alias{GANTrainer}
\title{GANTrainer}
\usage{
GANTrainer(
  data,
  noise_dim = 2,
  noise_distribution = "normal",
  value_function = "original",
  generator = NULL,
  generator_optimizer = NULL,
  discriminator = NULL,
  discriminator_optimizer = NULL,
  batch_size = 50,
  epochs = 150,
  plot = FALSE,
  device = "cpu"
)
}
\arguments{
\item{device}{Input on which device (e.g. "cpu" or "cuda") will you be training?}

\item{transformed_data}{Input a data set after DataTransformer}
}
\value{
A function
}
\description{
Provides a function to send the output of a DataTransformer to
a torch tensor, so that it can be accessed during GAN training.
}
