% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator-architectures.R
\name{Generator}
\alias{Generator}
\title{Generator}
\usage{
Generator(
  noise_dim,
  data_dim,
  hidden_units = list(128, 128),
  dropout_rate = 0.5
)
}
\arguments{
\item{noise_dim}{The length of the noise vector per example}

\item{data_dim}{The number of columns in the data set}

\item{hidden_units}{A list of the number of neurons per layer, the length of the list determines the number of hidden layers}

\item{dropout_rate}{The dropout rate for each hidden layer}
}
\value{
A torch::nn_module for the Generator
}
\description{
Provides a torch::nn_module with a simple fully connected neural
net, for use as the default architecture for tabular data in RGAN.
}
