% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator-architectures.R
\name{TabularGenerator}
\alias{TabularGenerator}
\title{Tabular Generator with Gumbel-Softmax}
\usage{
TabularGenerator(
  noise_dim,
  output_info,
  hidden_units = list(256, 256),
  dropout_rate = 0.5,
  tau = 0.2
)
}
\arguments{
\item{noise_dim}{The length of the noise vector per example}

\item{output_info}{A list describing the output structure from data_transformer$output_info.
Each element is a list with (dimension, type) where type is "linear", "mode_specific", or "softmax".}

\item{hidden_units}{A list of the number of neurons per layer}

\item{dropout_rate}{The dropout rate for each hidden layer}

\item{tau}{Temperature for Gumbel-Softmax. Lower values produce more discrete outputs. Defaults to 0.2.}
}
\value{
A torch::nn_module for the Tabular Generator
}
\description{
Provides a torch::nn_module Generator for tabular data that applies
Gumbel-Softmax to categorical outputs for differentiable sampling. This improves
gradient flow for discrete variables compared to standard softmax.
}
