% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value-functions.R
\name{WGAN_weight_clipper}
\alias{WGAN_weight_clipper}
\title{WGAN Weight Clipper}
\usage{
WGAN_weight_clipper(d_net, clip_values = c(-0.01, 0.01))
}
\arguments{
\item{d_net}{A torch::nn_module (typically a discriminator/critic) for which the weights should be clipped}

\item{clip_values}{A vector with the lower and upper bound for weight values. Any value outside this range will be set to the closer value.}
}
\value{
The function modifies the torch::nn_module weights in place
}
\description{
A function that clips the weights of a Discriminator (for WGAN training).
}
